+++ webapi/src/main/scala/org/knora/webapi/routing/ApiRoutes.scala
@@ -30,6 +30,7 @@ import org.knora.webapi.routing.v2._
 import org.knora.webapi.slice.admin.api.ProjectsEndpointsHandlerF
 import org.knora.webapi.slice.admin.api.service.ProjectADMRestService
 import org.knora.webapi.slice.admin.domain.service.KnoraProjectRepo
+import org.knora.webapi.slice.common.api.TapirToPekkoInterpreter
 import org.knora.webapi.slice.ontology.api.service.RestCardinalityService
 import org.knora.webapi.slice.resourceinfo.api.RestResourceInfoService
 import org.knora.webapi.slice.resourceinfo.domain.IriConverter
